+++ NEWS.md
@@ -1,5 +1,8 @@
 # permChacko (development version)
 
+* Improved validation
+* Optimized `chackoStatistic()`
+
 # permChacko 1.0.0
 
 * Improved printing of hypothesis ([issue #11](https://github.com/ocbe-uio/permChacko/issues/11)).